update structural demo
