--- client/src/components/login/login.js
@@ -1,134 +1,0 @@
-import React, { Component } from 'react';
-import { StyleSheet, Text, View, TextInput, TouchableOpacity } from 'react-native';
-import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
-import { login } from '../actions/authAction';
-import { clearErrors } from '../actions/errorActions';
-
-class Login extends Component {
-  state = {
-    modal: false,
-    email: 'nguyentuanvu231198@gmail.com',
-    password: '123456',
-    msg: null
-  };
-  //định nghĩa các prop
-  static propTypes = {
-    isAuthenticated: PropTypes.bool,
-    error: PropTypes.object.isRequired,
-    login: PropTypes.func.isRequired,
-    clearErrors: PropTypes.func.isRequired
-  };
-
-  onSubmit = e => {
-    e.preventDefault();//chấp nó bấm submit liên tục nè
-    const { email, password } = this.state;
-    const user = {
-      email,
-      password
-    };
-    // Attempt to login
-    this.props.login(user);
-
-  };
-  //lưu những thay đổi nơi input vào state
-  onChange = e => {
-    this.setState({ [e.target.name]: e.target.value });
-  };
-  
-  render() {
-    return (
-        <View style={styles.container}>
-            <Text style={styles.logo}>HeyAPP</Text>
-            <View style={styles.inputView} >
-                <TextInput
-                    id="login"
-                    name="email"
-                    style={styles.inputText}
-                    value={this.state.email}
-                    placeholder="Email..."
-                    placeholderTextColor="#003f5c"
-                    onChangeText={text => this.setState({ email: text })} 
-                    />
-            </View>
-            <View style={styles.inputView} >
-                <TextInput
-                    id="password" 
-                    name="password"
-                    value={this.state.password}
-                    secureTextEntry
-                    style={styles.inputText}
-                    placeholder="Password..."
-                    placeholderTextColor="#003f5c"
-                    onChangeText={text => this.setState({ password: text })} 
-                    />
-            </View>
-            <TouchableOpacity>
-                <Text style={styles.forgot}>Forgot Password?</Text>
-            </TouchableOpacity>
-            <TouchableOpacity style={styles.loginBtn} onPress={this.onSubmit}>
-                <Text style={styles.loginText}>LOGIN</Text>
-            </TouchableOpacity>
-        <TouchableOpacity onPress={() =>
-          this.props.navigation.navigate('register')}>
-                <Text style={styles.loginText}>Signup</Text>
-            </TouchableOpacity>
-        </View>
-    );
-  }
-}
-
-const mapStateToProps = state => ({
-  isAuthenticated: state.auth.isAuthenticated,
-  error: state.error
-});
-
-export default connect(
-  mapStateToProps,
-  { login, clearErrors }
-)(Login);
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        backgroundColor: '#003f5c',
-        alignItems: 'center',
-        justifyContent: 'center',
-    },
-    logo: {
-        fontWeight: "bold",
-        fontSize: 50,
-        color: "#fb5b5a",
-        marginBottom: 40
-    },
-    inputView: {
-        width: "80%",
-        backgroundColor: "#465881",
-        borderRadius: 25,
-        height: 50,
-        marginBottom: 20,
-        justifyContent: "center",
-        padding: 20
-    },
-    inputText: {
-        height: 50,
-        color: "white"
-    },
-    forgot: {
-        color: "white",
-        fontSize: 11
-    },
-    loginBtn: {
-        width: "80%",
-        backgroundColor: "#fb5b5a",
-        borderRadius: 25,
-        height: 50,
-        alignItems: "center",
-        justifyContent: "center",
-        marginTop: 40,
-        marginBottom: 10
-    },
-    loginText: {
-        color: "white"
-    }
-});
