--- client/src/components/screens/cameraScreen.js
@@ -1,11 +1,0 @@
-import React, {Component} from 'react';
-import {Platform, StyleSheet, Text, View, Button} from 'react-native';
-// import CameraRN from './Camera/RNCamera';
-// import ImagePickerScreen from './Camera/ImagePicker';
-export default function CameraScreen({navigation}) {
-  return (
-    <View style={{flex: 1, alignItems: 'center', justifyContent: 'center'}}>
-      {/* <ImagePickerScreen /> */}
-    </View>
-  );
-}
