+++ client/src/screens/HomeScreen.js
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import {View, Text} from 'react-native';
+import { styles } from '../public/styleSheets/styleViewInfo';
+
+export default function UserInfo({navigation}) {
+  return (
+    <View style={styles.container}>
+      <Text>Home Screen</Text>
+    </View>
+  );
+}
