+++ client/src/screens/UserScreen.js
@@ -1,0 +1,11 @@
+import * as React from 'react';
+import {View, Text} from 'react-native';
+import {styles} from '../public/styleSheets/styleViewInfo';
+
+export default function UserScreen({navigation}) {
+  return (
+    <View style={styles.container}>
+     <Text>User Screen</Text>
+    </View>
+  );
+}
